rent book project

+++ book-rental-project-v1/book-rental-backend/src/main/java/com/example/bookrentalbackend/service/BookService.java
@@ -5,8 +5,10 @@ import com.example.bookrentalbackend.util.FileHandler;
 import com.example.bookrentalbackend.vo.BookFileVO;
 import com.example.bookrentalbackend.vo.search.BookSearchVO;
 import com.example.bookrentalbackend.vo.BookVO;
+import com.example.bookrentalbackend.vo.search.RentalSearchVO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -24,31 +26,46 @@ public class BookService {
         return bookMapper.getAll();
     }
 
-    public List<BookVO> getBookList(BookSearchVO vo) {
-        return bookMapper.getBookList(vo);
+    public List<BookVO> getBookList(BookSearchVO bookSearchVO) {
+        return bookMapper.getBookList(bookSearchVO);
+    }
+
+    public int getBookCount(BookSearchVO bookSearchVO) {
+        return bookMapper.getBookCount(bookSearchVO);
     }
 
     public BookVO getBookDetail(long bookSeq) {
         BookVO bookVO = bookMapper.findBookBySeq(bookSeq);
-        List<BookFileVO> bookFileVOS = bookMapper.getBookFileList( Map.of("bookSeq", bookSeq) );
-        bookVO.setBookFiles(bookFileVOS);
+//        List<BookFileVO> bookFileVOS = bookMapper.getBookFileList( Map.of("bookSeq", bookSeq) );
+//        bookVO.setBookFiles(bookFileVOS);
 
         return bookVO;
     }
 
+    public List<BookFileVO> getBookFileList(long bookSeq) {
+        return bookMapper.getBookFileList( Map.of("bookSeq", bookSeq) );
+    }
+
     public void insertBook(BookVO bookVO, List<MultipartFile> files) throws IOException {
         bookMapper.insertBook(bookVO);
-        if (bookVO.getSeq() > 0) {
+        if (files != null && bookVO.getSeq() > 0) {
             _storeFiles(bookVO.getSeq(), files);
         }
     }
 
-    public void updateBook(BookVO bookVO, List<MultipartFile> files, List<Long> deleteFileSeqs) throws IOException {
+    public void updateBook(BookVO bookVO) throws IOException {
         bookMapper.updateBook(bookVO);
-        _storeFiles(bookVO.getSeq(), files);
+    }
 
-        List<BookFileVO> bookFileVOS = bookMapper.getBookFileList( Map.of("seqs", deleteFileSeqs) );
-        _deleteFiles(bookFileVOS);
+    public void updateBookFile(long bookSeq, List<MultipartFile> files, List<Long> deleteFileSeqs) throws IOException {
+        if ( files != null ) {
+            _storeFiles(bookSeq, files);
+        }
+
+        if( !CollectionUtils.isEmpty(deleteFileSeqs) ) {
+            List<BookFileVO> bookFileVOS = bookMapper.getBookFileList( Map.of("seqs", deleteFileSeqs) );
+            _deleteFiles(bookFileVOS);
+        }
     }
 
     public void deleteBook(long bookSeq) {
